+++ source/Application/Controller/BasketController.php
@@ -105,7 +105,8 @@ class BasketController extends \OxidEsales\Eshop\Application\Controller\Frontend
     public function render()
     {
         if (\OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('blPsBasketReservationEnabled')) {
-            $this->getSession()->getBasketReservations()->renewExpiration();
+            $session = \OxidEsales\Eshop\Core\Registry::getSession();
+            $session->getBasketReservations()->renewExpiration();
         }
 
         parent::render();
@@ -124,7 +125,8 @@ class BasketController extends \OxidEsales\Eshop\Application\Controller\Frontend
             $this->_oBasketArticles = false;
 
             // passing basket articles
-            if ($oBasket = $this->getSession()->getBasket()) {
+            $session = \OxidEsales\Eshop\Core\Registry::getSession();
+            if ($oBasket = $session->getBasket()) {
                 $this->_oBasketArticles = $oBasket->getBasketArticles();
             }
         }
@@ -214,7 +216,8 @@ class BasketController extends \OxidEsales\Eshop\Application\Controller\Frontend
             return;
         }
 
-        $oBasket = $this->getSession()->getBasket();
+        $session = \OxidEsales\Eshop\Core\Registry::getSession();
+        $oBasket = $session->getBasket();
         $oBasket->addVoucher(\OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter('voucherNr'));
     }
 
@@ -229,7 +232,8 @@ class BasketController extends \OxidEsales\Eshop\Application\Controller\Frontend
             return;
         }
 
-        $oBasket = $this->getSession()->getBasket();
+        $session = \OxidEsales\Eshop\Core\Registry::getSession();
+        $oBasket = $session->getBasket();
         $oBasket->removeVoucher(\OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter('voucherId'));
     }
 
@@ -335,7 +339,8 @@ class BasketController extends \OxidEsales\Eshop\Application\Controller\Frontend
         $oConfig = \OxidEsales\Eshop\Core\Registry::getConfig();
 
         if ($this->getViewConfig()->getShowGiftWrapping()) {
-            $oBasket = $this->getSession()->getBasket();
+            $session = \OxidEsales\Eshop\Core\Registry::getSession();
+            $oBasket = $session->getBasket();
 
             $this->_setWrappingInfo($oBasket, $oConfig->getRequestParameter('wrapping'));
 
@@ -370,8 +375,10 @@ class BasketController extends \OxidEsales\Eshop\Application\Controller\Frontend
      */
     public function getBasketContentMarkGenerator()
     {
+        $session = \OxidEsales\Eshop\Core\Registry::getSession();
+
         /** @var \OxidEsales\Eshop\Application\Model\BasketContentMarkGenerator $oBasketContentMarkGenerator */
-        return oxNew('oxBasketContentMarkGenerator', $this->getSession()->getBasket());
+        return oxNew('oxBasketContentMarkGenerator', $session->getBasket());
     }
 
     /**