+++ source/Application/Model/Article.php
@@ -1070,7 +1070,8 @@ class Article extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel implements
         if (\OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('blUseStock') && $this->oxarticles__oxstockflag->value == 2) {
             $iOnStock = $this->oxarticles__oxstock->value + $this->oxarticles__oxvarstock->value;
             if (\OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('blPsBasketReservationEnabled')) {
-                $iOnStock += $this->getSession()->getBasketReservations()->getReservedAmount($this->getId());
+                $session = \OxidEsales\Eshop\Core\Registry::getSession();
+                $iOnStock += $session->getBasketReservations()->getReservedAmount($this->getId());
             }
             if ($iOnStock <= 0) {
                 return false;
@@ -2491,7 +2492,8 @@ class Article extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel implements
             }
         }
         if (\OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('blPsBasketReservationEnabled')) {
-            $iOnStock += $this->getSession()->getBasketReservations()->getReservedAmount($this->getId());
+            $session = \OxidEsales\Eshop\Core\Registry::getSession();
+            $iOnStock += $session->getBasketReservations()->getReservedAmount($this->getId());
         }
         if ($iOnStock >= $dAmount) {
             return true;
@@ -4404,7 +4406,10 @@ class Article extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel implements
         if ($myConfig->getConfigParam('blUseStock') && ($this->oxarticles__oxstockflag->value == 3 || $this->oxarticles__oxstockflag->value == 2)) {
             $iOnStock = $this->oxarticles__oxstock->value;
             if (\OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('blPsBasketReservationEnabled')) {
-                $iOnStock += $this->getSession()->getBasketReservations()->getReservedAmount($this->getId());
+                $session = \OxidEsales\Eshop\Core\Registry::getSession();
+                if ($reservations = $session->getBasketReservations()){
+                    $iOnStock += $reservations->getReservedAmount($this->getId());
+                }
             }
             if ($iOnStock <= 0) {
                 $this->setBuyableState(false);