+++ tests/Unit/Application/Controller/Admin/DeliverySetListTest.php
@@ -23,12 +23,11 @@ class DeliverySetListTest extends \OxidTestCase
         oxTestModules::addFunction("oxUtilsServer", "getOxCookie", "{return array(1);}");
         oxTestModules::addFunction("oxUtils", "checkAccessRights", "{return true;}");
 
-        $oSess = $this->getMock(\OxidEsales\Eshop\Core\Session::class, array('checkSessionChallenge'));
-        $oSess->expects($this->any())->method('checkSessionChallenge')->will($this->returnValue(true));
+        $session = $this->getMock(\OxidEsales\Eshop\Core\Session::class, array('checkSessionChallenge'));
+        $session->expects($this->any())->method('checkSessionChallenge')->will($this->returnValue(true));
+        \OxidEsales\Eshop\Core\Registry::set(\OxidEsales\Eshop\Core\Session::class, $session);
 
-        $oView = $this->getMock($this->getProxyClassName('DeliverySet_List'), array('getSession'));
-        $oView->expects($this->any())->method('getSession')->will($this->returnValue($oSess));
-
+        $oView = oxNew($this->getProxyClassName('DeliverySet_List'));
         $oView->init();
 
         $this->assertEquals('oxdeliveryset', $oView->getNonPublicVar("_sListClass"));